backend/requirements.txt
Django>=2.1.3,<2.2.0
uWSGI>=2.0.18,<2.1
psycopg2-binary==2.8.3
djangorestframework>=3.9.0,<3.10.0
Pillow>=5.3.0, <5.4.0
flake8>=3.6.0,<3.7.0
django-cors-headers<=3.8.0
django-storages<=1.11.1
boto
boto3